Nested structs work, but not for auto-complete/suggestions
It is now possible to call elements inside structs.
If element does not exist we get red underline, and if otherwise it shows ok.
Hover also works for all the items.
HOWEVER, the autocomplete suggestions do not works.
On first `.` it just suggests top level items again. And on deeper levels, it just shows the VS Code *abc*.