Fix `NUMBER` terminal consuming trailing + or -
If a number was entered with a trailing `-` or `+` with no spaces
between it, the NUMBER terminal consumed it. This meant that further
parsing failed since there should have been an addition symbol was
expected (but NUMBER terminal had incorrectly consumed it).
This commit makes the NUMBER terminal a bit more strict, requiring that
a number must follow `e`, `+` or `-` etc.

Add .scl file demonstrating issue NUMBER terminal consuming trailing addition