Fix broken region and titles when they start with keyword
This is a workaround for the issue where parsing/lexing of a region would
fail if the region text started with a word that contained a keyword.
E.g. `Interval ...` would cause it to fail, since `INT` is a keyword
(and we have switched off case sensitivity).
Note that `Some Interval ...` would not cause it to fail, since it was
only an issue if it was the start of the terminal.

This commit makes `REGION` and `TITLE = ` part of the terminal, instead
of using positive lookbehind.