Restore optional `#` when referencing MemberCall
This was removed in 0ece5f5dd, but it turns out that it is actually
possible to use `#` in a member call. TIA Portal will not give auto
complete when starting with it though, so it is not the normal way to
program. But it will not error out if you enter it manually, so it is
best to support it.