Move declarations into Declaration Blocs
Benefits:
We now know, from container, if it is input, or output etc.
Also, we get same structure for UDTs and declaration sections.

Disadvantage:
EMPTY_SCOPE no longer returns all declarations in block, we need to do some computation for them. But performance hit may be negligible.

Current status. Some more complex test file is failing to resolve referenced due to someting in the program section. Investigation to be performed.