Remove global scope override
By default top level elements in AST is accessible every where.
Originally global scope was overridden to return EMPTY_SCOPE, to avoid
local variables from other blocks showing up in current block. And then,
to enable global block names (UDTs) to be used in variable declarations,
the global scope override was extended to return those in that situation.

But the VariableDeclarations have since been moved into VarDec blocks, so
that the override is no longer needed for this. Therefore the override
can be removed, and same functionality be retained.