Add SclBlocks to getScope (for FeatureCall/MemberCall)
These were only available from global scope when accessed via SclBlock
and not when accessed via NamedElement. Hence they would only resolve
for VariableDeclarations (which uses SclBlock) and not for FeatureCall
(which uses NamedElement).
This commit adds SclBock nodes to the `getScope` for !previous, so that
it will also resolve these there.