Override `isAffected` but still use `super.isAffected`
Add framework for override so that we can always return false, or create
custom rules for when document is considered changed.

Just returning false will prevent recalculating broken references when
updating a file. Renaming linked objects with rename functionality will
still work, but if you just edit something directly, it the link will
remain to an item which now has a different name.

For now, the isAffected function just returns the normal isAffected.