Add navigate to the builtin library element in VS Code
This registers an in-memory file system for vscode to use for the builtin file schema. Every time vscode is supposed to open a file with this schema, it will invoke the stat and readFile methods of the registered file system provider.

Note:
In this version it always returns the same library file.
TODO: Return file for relevant URI.