Add uriMap for builtin SCL library functions for easier maintenance
This makes it easier to add new functions over time, since the
*scl-workspace-manager.ts* and *scl-library-file-system-provider.ts* will
now use a common map defined in *built-in-scl-library-functions.ts* to
register the libraries as documents, and to provide correct document for
navigation in VS Code.

With this we avoid make changes in three different locations when adding
new function.