Add input parameter to built in library convert functions
In SCL, if a function only has a single formal parameter in INPUT, OUTPUT
or IN_OUT, we can omit it when calling the function and just input the
argument directly.

This commit adds an `IN` parameter to the builtin convert functions, so
that it will handle both if that is included and not.