Add caching for local and global scope
This significantly improves performance, since data can be collected from
cache.

The local cache will get invalidated, and therefore recalculated
it the document with stored URI changes. But in addition any document
with broken links will also get recalculated and therefore not gain the
full benefit of caching if you have many documents with broken links.

The global cache gets invalidated for any change to the workspace.

Updated URI for one of the tests so that it is unique, since the test,
unlike when used in VS Code, does not invalidate the cache.