Add includeFolder filter possibility
Can read file in root of workspace that contains list of folders to be
included. If file does not exist, all folders gets included. Note that
that files directly in root of working directory will still get included
regardless of filter config.

The filter is only applied for the initial loading. If files that are not
included in the filter gets opened in VS Code, that they will get loaded
into the LSP by VS Code, even if they are not included in the filter list.