Narrow down xml support for tags to `tags.xml`
The `UriUtils.extname(uri)` returns only `.xml` for `.tags.xml` so
hardcoded support for `.tags.xml` into it.
Update language to target `.tags.xml` instead of `.xml`. So now only
tag files are included in the LSP.

Known issue:
When extension is triggering it does not parse/link the `.tags.xml` files.
Linking is resolved to a specific tags file only after opening it.