Parser: Add parsing of nested multi-line comments
It is not possible to handle nested multi-line comments with regex.
This commit implements a *MultiModeLexer* by creating my own class `SclTokenBuilder` overriding functions from `DefaultTokenBuilder`.

Change `(` from a keyword to a terminal in the grammar, and place it after the `(*` terminal.
This prevents the *LONGER_ALT* issue.
Also add `*` as terminal. But we could have kept that as a keyword, and instead deleted the *LONGER_ALT* from it in the `builtKeywordToken()` function override.
The reason we could not do that for `(` is that we would then be unable to make the `(*` terminal match anything, since we would have deleted that as the *LONGER_ALT*.

Useful sources:
- [docs under construction - Guide on multi-mode lexing #132](https://github.com/eclipse-langium/langium-website/pull/132/files)
- [Q&A - Answer to subquestion regarding nested mulit-line comments in: "Discussion: Tooling Development for an Existing DSL #609" ](https://github.com/eclipse-langium/langium/discussions/609#discussioncomment-3229263) This one is from summer 2022 before lexer modes were released).