Parser XML: Add support for comment starting with `=`
Hardcoded support for a single `=` at the beginning of a tag comment.
Previously this was not supported since it is a keyword, which gets
evaluated before terminals. Hence is did not become part of the
"TAG_COMMENT" terminal.

If necessary we can make a better fix later with lexer modes.