Built-ins: Fix data types used in T_CONV functions
Sone of the functions used from/to names that was not a valid data-type.
Replace with correct type, and feed in function name separately.

The "type" used in function name for these is actually the shorted prefix
variant used for literals. So for some of the functions they use the type
as name, and for some the literal prefix.